+++ phpmyfaq/src/phpMyFAQ/Instance/Database/Mysqli.php
@@ -291,27 +291,6 @@ class Mysqli extends Database implements Driver
             searchdate TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
             PRIMARY KEY (id, lang)) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci ENGINE = InnoDB',
 
-        'faqsections' => 'CREATE TABLE %sfaqsections (
-            id INT(11) NOT NULL,
-            name VARCHAR(255) NOT NULL,
-            description VARCHAR(255) DEFAULT NULL,
-            PRIMARY KEY (id)) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci ENGINE = InnoDB',
-
-        'faqsection_category' => 'CREATE TABLE %sfaqsection_category (
-            section_id INT(11) NOT NULL,
-            category_id INT(11) NOT NULL DEFAULT -1,
-            PRIMARY KEY (section_id, category_id))',
-
-        'faqsection_group' => 'CREATE TABLE %sfaqsection_group (
-            section_id INT(11) NOT NULL,
-            group_id INT(11) NOT NULL DEFAULT -1,
-            PRIMARY KEY (section_id, group_id))',
-
-        'faqsection_news' => 'CREATE TABLE %sfaqsection_news (
-            section_id INT(11) NOT NULL,
-            news_id INT(11) NOT NULL DEFAULT -1,
-            PRIMARY KEY (section_id, news_id))',
-
         'faqsessions' => 'CREATE TABLE %sfaqsessions (
             sid INT(11) NOT NULL,
             user_id INT(11) NOT NULL,