+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 - added new mobile first, touch-friendly default layout (Thorsten)
 - added HTTPS support for local Docker development (Thorsten)
 - added Monolog v2 as logging solution (Thorsten)
+- removed experimental large permissions support with sections (Thorsten)
 - migrated from SwiftMailer to Symfony Mailer (Thorsten)
 - migrated codebase to use PHP 8.0 language features
 - updated to Bootstrap v5.3 (Thorsten)